pyproject.toml
[project]
name = "infohub-chatbot-langchain"
version = "0.1.0"
description = "InfoHub Chatbot project managed with uv"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "openai>=1.40,<2.0",
  "beautifulsoup4>=4.12,<5.0",
  "pandas>=2.2,<3.0",
  "scipy>=1.13,<2.0",
  "tiktoken>=0.7,<1.0",
  "fastapi>=0.116,<1.0",
  "uvicorn>=0.30,<1.0",
  "redis>=5.0,<6.0",
  "psycopg2-binary>=2.9,<3.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.uv]
package = false

